+++ src/main/java/no/ntnu/vislab/vislabcontroller/repositories/DeviceTypeRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import no.ntnu.vislab.vislabcontroller.entity.DeviceType;
 
-public interface DeviceTypeRepository extends JpaRepository<DeviceType, Long> {
+public interface DeviceTypeRepository extends JpaRepository<DeviceType, Integer> {
 }
